# docs/swagger.yaml
basePath: /
definitions:
...
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: Reusable Go REST API backend template.
  title: WarSpark API
  version: dev
paths:
...
  /api/v1/leagues:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取联赛列表
      tags:
      - league
  /api/v1/leagues/{id}:
    get:
      parameters:
      - description: 联赛ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取联赛详情
      tags:
      - league
  /api/v1/leagues/{id}/seasons:
    get:
      parameters:
      - description: 联赛ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取联赛赛季列表
      tags:
      - league
  /api/v1/leagues/{id}/seasons/{season}/rankings:
    get:
      parameters:
      - description: 联赛ID
        in: path
        name: id
        required: true
        type: string
      - description: 赛季ID
        in: path
        name: season
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取联赛赛季排名
      tags:
      - league
  /api/v1/leagues/{id}/seasons/{season}/tiers:
    get:
      parameters:
      - description: 联赛ID
        in: path
        name: id
        required: true
        type: string
      - description: 赛季ID
        in: path
        name: season
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取联赛赛季段位列表
      tags:
      - league
  /api/v1/leagues/{id}/seasons/{season}/tiers/{tier}:
    get:
      parameters:
      - description: 段位ID
        in: path
        name: tier
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取联赛赛季段位详情
      tags:
      - league
  /api/v1/leagues/{id}/seasons/{season}/tiers/{tier}/history:
    get:
      parameters:
      - description: 玩家标签
        in: query
        name: player_tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取玩家联赛赛季历史
      tags:
      - league
...
  /api/v1/players/{tag}:
    get:
      parameters:
      - description: 玩家标签
        in: path
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取玩家概览
      tags:
      - player
  /api/v1/players/{tag}/battle-log:
    get:
      parameters:
      - description: 玩家标签
        in: path
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取玩家战斗日志
      tags:
      - player
...
  /api/v1/war-leagues:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取战争联赛列表
      tags:
      - league
  /api/v1/war-leagues/{id}:
    get:
      parameters:
      - description: 联赛ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取战争联赛详情
      tags:
      - league
...
schemes:
- http
swagger: "2.0"